added link functionality in nav bar, moved Navbar out of pages/home, so that it can be outside of <Routes>
