+++ src/pages/Home/Home.css
@@ -84,19 +84,22 @@
 
 /*logo */
 .img {
+  top: 0px;
   width: 9vw;
   height: auto;
   position: fixed;
   transition: transform 0.8s ease;
 }
+
 .img:hover {
   transform: translateY(10px);
 }
+
 /* NavBar*/
 .NavBar {
   font-family: "Courier New", monospace;
   font-size: 1.6vw;
-
+  top: 0px;
   position: fixed;
   right: 16vw;
   list-style-type: none;
@@ -115,6 +118,7 @@
     0 0 10px rgba(189, 189, 177, 0.923), 0 0 30px rgba(99, 98, 96, 0.227);
   transform: translateY(-10px);
 }
+
 .subpage::after {
   content: "";
   position: absolute;
@@ -131,6 +135,7 @@
 .subpage:hover::after {
   transform: scaleX(1);
 }
+
 /*bottom*/
 .Bottom {
   position: absolute;
@@ -139,14 +144,13 @@
   top: 400vh;
   background-color: rgba(0, 0, 0, 0.83);
 }
+
 /*top*/
 .Top {
-  background: linear-gradient(
-    to bottom,
-    rgba(249, 236, 170, 1) 0%,
-    rgba(249, 236, 170, 0.6) 50%,
-    rgba(249, 236, 170, 0) 100%
-  );
+  background: linear-gradient(to bottom,
+      rgba(249, 236, 170, 1) 0%,
+      rgba(249, 236, 170, 0.6) 50%,
+      rgba(249, 236, 170, 0) 100%);
   position: fixed;
   top: 0px;
   left: 0px;
